move CriticalSpring to its own file
